# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/assignments": {
      "get": {
        "operationId": "AssignmentsController_list_api/v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields; prefix with - for descending, e.g. -createdAt,lastName",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "homework",
                "quiz",
                "test",
                "project",
                "essay",
                "presentation",
                "lab",
                "discussion",
                "practice"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "closed"
              ]
            }
          },
          {
            "name": "dueBefore",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dueAfter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List assignments (students and parents: published work in their classes, with their own status)",
        "tags": [
          "Assignments"
        ]
      },
      "post": {
        "operationId": "AssignmentsController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft assignment in a class you manage",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/assignments/{id}": {
      "get": {
        "operationId": "AssignmentsController_get_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assignment detail (students: with their own submissions)",
        "tags": [
          "Assignments"
        ]
      },
      "patch": {
        "operationId": "AssignmentsController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an assignment",
        "tags": [
          "Assignments"
        ]
      },
      "delete": {
        "operationId": "AssignmentsController_remove_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an ungraded assignment",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/assignments/{id}/publish": {
      "post": {
        "operationId": "AssignmentsController_publish_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish to students",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/assignments/{id}/close": {
      "post": {
        "operationId": "AssignmentsController_close_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close: no further submissions",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/assignments/{id}/submissions": {
      "get": {
        "operationId": "AssignmentsController_submissions_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Every enrolled student with their latest submission and grade",
        "tags": [
          "Assignments"
        ]
      },
      "post": {
        "operationId": "AssignmentsController_submit_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit text and/or uploaded file ids; the attempt number is assigned",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/submissions/{id}": {
      "get": {
        "operationId": "SubmissionsController_get_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One submission (the student, their guardians, or staff managing the class)",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/submissions/{id}/grade": {
      "post": {
        "operationId": "SubmissionsController_grade_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grade a submission: score, feedback, rubric scores; late penalty applied unless waived",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/grades": {
      "get": {
        "operationId": "GradesController_list_api/v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields; prefix with - for descending, e.g. -createdAt,lastName",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grades: all in scope for staff, own for students, children for parents",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/rubrics": {
      "get": {
        "operationId": "RubricsController_list_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rubrics in the organisation",
        "tags": [
          "Assignments"
        ]
      },
      "post": {
        "operationId": "RubricsController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRubricDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a rubric",
        "tags": [
          "Assignments"
        ]
      }
    },
    "/api/v1/rubrics/{id}": {
      "get": {
        "operationId": "RubricsController_get_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One rubric",
        "tags": [
          "Assignments"
        ]
      },
      "patch": {
        "operationId": "RubricsController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRubricDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a rubric",
        "tags": [
          "Assignments"
        ]
      },
      "delete": {
        "operationId": "RubricsController_remove_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a rubric",
        "tags": [
          "Assignments"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartSchool API",
    "description": "LMS API for SmartSchool. Conventions: docs/09-API-DESIGN.md. Errors use RFC 9457 problem details.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAssignmentDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "homework",
              "quiz",
              "test",
              "project",
              "essay",
              "presentation",
              "lab",
              "discussion",
              "practice"
            ],
            "default": "homework"
          },
          "submissionType": {
            "type": "string",
            "enum": [
              "online",
              "paper",
              "in_person",
              "external",
              "no_submission"
            ],
            "default": "online"
          },
          "category": {
            "type": "string",
            "example": "Homework"
          },
          "maxPoints": {
            "type": "number",
            "default": 100
          },
          "weight": {
            "type": "number",
            "default": 1
          },
          "availableFrom": {
            "type": "string"
          },
          "dueAt": {
            "type": "string"
          },
          "allowLateUntil": {
            "type": "string"
          },
          "latePenaltyPercent": {
            "type": "number",
            "example": 10
          },
          "maxAttempts": {
            "type": "number",
            "example": 3
          },
          "rubricId": {
            "type": "string",
            "format": "uuid"
          },
          "h5pContentId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "A grade category of the class; sets the category label"
          },
          "isExtraCredit": {
            "type": "boolean",
            "default": false,
            "description": "Adds to the score without adding to the possible points"
          },
          "gradingPeriodId": {
            "type": "string",
            "format": "uuid"
          },
          "standardIds": {
            "description": "Standards this work assesses",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "classId",
          "title"
        ]
      },
      "UpdateAssignmentDto": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "homework",
              "quiz",
              "test",
              "project",
              "essay",
              "presentation",
              "lab",
              "discussion",
              "practice"
            ],
            "default": "homework"
          },
          "submissionType": {
            "type": "string",
            "enum": [
              "online",
              "paper",
              "in_person",
              "external",
              "no_submission"
            ],
            "default": "online"
          },
          "category": {
            "type": "string",
            "example": "Homework"
          },
          "maxPoints": {
            "type": "number",
            "default": 100
          },
          "weight": {
            "type": "number",
            "default": 1
          },
          "availableFrom": {
            "type": "string"
          },
          "dueAt": {
            "type": "string"
          },
          "allowLateUntil": {
            "type": "string"
          },
          "latePenaltyPercent": {
            "type": "number",
            "example": 10
          },
          "maxAttempts": {
            "type": "number",
            "example": 3
          },
          "rubricId": {
            "type": "string",
            "format": "uuid"
          },
          "h5pContentId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "A grade category of the class; sets the category label"
          },
          "isExtraCredit": {
            "type": "boolean",
            "default": false,
            "description": "Adds to the score without adding to the possible points"
          },
          "gradingPeriodId": {
            "type": "string",
            "format": "uuid"
          },
          "standardIds": {
            "description": "Standards this work assesses",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "SubmitDto": {
        "type": "object",
        "properties": {
          "textContent": {
            "type": "string"
          },
          "fileIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "RubricScoreDto": {
        "type": "object",
        "properties": {
          "criterionId": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "criterionId",
          "points"
        ]
      },
      "StandardScoreDto": {
        "type": "object",
        "properties": {
          "standardId": {
            "type": "string",
            "format": "uuid"
          },
          "level": {
            "type": "number",
            "example": 3,
            "description": "A level of the class proficiency scale"
          }
        },
        "required": [
          "standardId",
          "level"
        ]
      },
      "GradeSubmissionDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "description": "Raw score before any late penalty"
          },
          "feedback": {
            "type": "string"
          },
          "rubricScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RubricScoreDto"
            }
          },
          "waiveLatePenalty": {
            "type": "boolean",
            "default": false,
            "description": "Skip the automatic late penalty"
          },
          "standardScores": {
            "description": "Standards-based classes: the level reached per standard; derived from the percentage when omitted",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StandardScoreDto"
            }
          }
        },
        "required": [
          "score"
        ]
      },
      "RubricLevelDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "label",
          "points"
        ]
      },
      "RubricCriterionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable id; generated when omitted"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "maxPoints": {
            "type": "number"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RubricLevelDto"
            }
          }
        },
        "required": [
          "title",
          "maxPoints"
        ]
      },
      "CreateRubricDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "criteria": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RubricCriterionDto"
            }
          },
          "isTemplate": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "title",
          "criteria"
        ]
      },
      "UpdateRubricDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "criteria": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RubricCriterionDto"
            }
          },
          "isTemplate": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  }
}